Move PhysicsSceneInstaller to Exanite.Arpg
Other changes:
Add documentation to PhysicsSceneInstaller.
Update LogInstaller documentation to be consistent with PhysicsSceneInstaller.
