+++ Assets/Exanite.Arpg/Installers/PhysicsSceneInstaller.cs
@@ -3,12 +3,18 @@ using UnityEngine;
 using UnityEngine.SceneManagement;
 using Zenject;
 
-namespace Prototype.DarkRift.Shared
+namespace Exanite.Arpg.Installers
 {
+    /// <summary>
+    /// Installs the <see cref="Scene"/> and <see cref="PhysicsScene"/> this component is in to a <see cref="DiContainer"/>
+    /// </summary>
     public class PhysicsSceneInstaller : MonoInstaller
     {
         [SerializeField] private bool requireLocalPhysicsScene = true;
 
+        /// <summary>
+        /// Should this <see cref="PhysicsSceneInstaller"/> require that this scene is loaded with a local <see cref="PhysicsScene"/>
+        /// </summary>
         public bool RequireLocalPhysicsScene
         {
             get
@@ -22,17 +28,26 @@ namespace Prototype.DarkRift.Shared
             }
         }
 
+        /// <summary>
+        /// Installs bindings to the <see cref="DiContainer"/>
+        /// </summary>
         public override void InstallBindings()
         {
             Container.Bind<Scene>().FromMethod(GetScene).AsSingle().NonLazy();
             Container.Bind<PhysicsScene>().FromMethod(GetPhysicsScene).AsSingle().NonLazy();
         }
 
+        /// <summary>
+        /// Gets the <see cref="Scene"/> this component is currently in
+        /// </summary>
         private Scene GetScene()
         {
             return gameObject.scene;
         }
 
+        /// <summary>
+        /// Gets the <see cref="PhysicsScene"/> this component is currently in
+        /// </summary>
         private PhysicsScene GetPhysicsScene()
         {
             var physicsScene = gameObject.scene.GetPhysicsScene();
@@ -44,5 +59,5 @@ namespace Prototype.DarkRift.Shared
 
             return physicsScene;
         }
-    } 
+    }
 }
